[리액트] key error 및 warning 제거
